Update README.md
As a student of the IB Diploma, this was the project I coded in Java for my SL Computer Science class which contributed to 30% of my total grade. I ended up receiving a 7 (A).

The idea behind the project is being able to predict the prices of ride-hailing apps (Uber/Grab/Lyft) through recreating their 'price surge' algorithm with the weather, distance, and peak hours being the main impacts on price.